backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
PyJWT[crypto]==2.10.1
cryptography==44.0.0
httpx==0.28.1
openai==2.33.0
python-multipart==0.0.19
supabase==2.11.0
langsmith==0.3.45
